Fix another circular dependency error in search module.
